생성 시 글의 속성은 일반만 가져야 한다 (#134)
* refactor: 엔티티 파일 위치 변경

* refactor: 엔티티 파일 위치 변경

* feat: post 속성 제거 및 게시판 식별자 추가

* refactor: 엔티티 추출 후 변경점

* feat: 게시글 생성을 위한 로직 구현

* test: 게시글 생성을 위한 테스트 코드 작성
