+++ databases.py
@@ -1,8 +1,8 @@
 from config.db_config import SessionLocal, Base, engine
 
-from app.models.comment_entity import CommentEntity
-from app.models.post_entity import PostEntity
-from app.models.user_entity import UserEntity
+from app.entity.comment_entity import CommentEntity
+from app.entity.post_entity import PostEntity
+from app.entity.user_entity import UserEntity
 
 
 def get_db() -> SessionLocal:
